+++ client/jsconfig.json
@@ -5,7 +5,7 @@
     "baseUrl": "./",
     "moduleResolution": "node",
     "paths": {
-      "@/*": [
+      "@": [
         "src/*"
       ]
     },
