+++ client/package.json
@@ -37,6 +37,7 @@
 		"pdap-design-system": "^2.5.0",
 		"postcss": "^8.4.32",
 		"tailwindcss": "^3.3.6",
+		"unplugin-vue-router": "^0.10.2",
 		"vite": "^4.5.1",
 		"vite-svg-loader": "^5.1.0",
 		"vitest": "^1.2.1"
