+++ client/src/pages/[...path].vue
@@ -1,5 +1,5 @@
 <template>
-	<main class="p-4">
+	<main class="pdap-flex-container py-12">
 		<h1>Something went wrong on our end</h1>
 		<p>
 			If you keep seeing this message, please email
