+++ client/src/pages/index.vue
